refactor: move raw input into input capture thread
