+++ crates/input-capture/Cargo.toml
@@ -14,6 +14,7 @@ windows = { workspace = true, features = [
     "Win32_Graphics_Gdi",
 ] }
 color-eyre.workspace = true
+tokio.workspace = true
 
 [dev-dependencies]
 tracing-subscriber.workspace = true
