+++ crates/input-capture/examples/input-read.rs
@@ -1,13 +1,14 @@
-use color_eyre::{Result, eyre::WrapErr as _};
-use input_capture::RawInput;
+use color_eyre::Result;
+use input_capture::InputCapture;
 
 pub fn main() -> Result<()> {
     color_eyre::install()?;
     tracing_subscriber::fmt::init();
 
-    let _raw_input = RawInput::initialize().expect("Failed to initialize raw input");
-    RawInput::run_queue(|event| {
+    let (_input_capture, mut input_rx) = InputCapture::new()?;
+    while let Some(event) = input_rx.blocking_recv() {
         tracing::info!(?event, "Received raw input event");
-    })
-    .wrap_err("failed to run message queue")
+    }
+
+    Ok(())
 }
