+++ crates/owl-recorder/src/main.rs
@@ -18,11 +18,8 @@ use clap::Parser;
 use color_eyre::{Result, eyre::eyre};
 
 use game_process::does_process_exist;
-use input_capture::{PressState, RawInput};
-use tokio::{
-    sync::{mpsc, oneshot},
-    time::MissedTickBehavior,
-};
+use input_capture::InputCapture;
+use tokio::{sync::oneshot, time::MissedTickBehavior};
 
 use crate::{
     idle::IdlenessTracker, keycode::lookup_keycode, raw_input_debouncer::EventDebouncer,
@@ -71,7 +68,7 @@ async fn main() -> Result<()> {
         )
     });
 
-    let mut input_rx = listen_for_raw_inputs();
+    let (_input_capture, mut input_rx) = InputCapture::new()?;
 
     let mut stop_rx = wait_for_ctrl_c();
 
@@ -81,6 +78,8 @@ async fn main() -> Result<()> {
     let mut perform_checks = tokio::time::interval(Duration::from_secs(1));
     perform_checks.set_missed_tick_behavior(MissedTickBehavior::Delay);
 
+    let mut debouncer = EventDebouncer::new();
+
     loop {
         tokio::select! {
             r = &mut stop_rx => {
@@ -89,8 +88,12 @@ async fn main() -> Result<()> {
             },
             e = input_rx.recv() => {
                 let e = e.expect("raw input reader was closed early");
+                if !debouncer.debounce(e) {
+                    continue;
+                }
+
                 recorder.seen_input(e).await?;
-                if let Some(key) = keycode_from_event(&e) {
+                if let Some(key) = e.key_press_keycode() {
                     if key == start_key {
                         tracing::info!("Start key pressed, starting recording");
                         recorder.start().await?;
@@ -129,40 +132,6 @@ async fn main() -> Result<()> {
     recorder.stop().await?;
 
     Ok(())
-}
-
-fn keycode_from_event(event: &input_capture::Event) -> Option<u16> {
-    if let input_capture::Event::KeyPress {
-        key,
-        press_state: PressState::Pressed,
-    } = event
-    {
-        Some(*key)
-    } else {
-        None
-    }
-}
-
-fn listen_for_raw_inputs() -> mpsc::Receiver<input_capture::Event> {
-    let (input_tx, input_rx) = mpsc::channel(1);
-
-    std::thread::spawn(move || {
-        let mut input_capture =
-            Some(RawInput::initialize().expect("raw input failed to initialize"));
-        let mut debouncer = EventDebouncer::new();
-
-        RawInput::run_queue(|event| {
-            if !debouncer.debounce(event) {
-                return;
-            }
-            if input_tx.blocking_send(event).is_err() {
-                tracing::debug!("Input channel closed, stopping raw input listener");
-                input_capture.take();
-            }
-        })
-        .expect("failed to run windows message queue");
-    });
-    input_rx
 }
 
 fn wait_for_ctrl_c() -> oneshot::Receiver<()> {
